+++ packages/webview_flutter/webview_flutter_android/example/lib/main.dart
@@ -148,9 +148,9 @@ Page resource error:
       ..loadRequest(LoadRequestParams(
         uri: Uri.parse('https://flutter.dev'),
       ))
-      ..setOnContentOffsetChanged((ContentOffsetChange contentOffsetChange) {
+      ..setOnScrollPositionChange((ScrollPositionChange scrollPositionChange) {
         debugPrint(
-            'Scroll offset change to x = ${contentOffsetChange.x}, y = ${contentOffsetChange.y}');
+            'Scroll offset change to x = ${scrollPositionChange.x}, y = ${scrollPositionChange.y}');
       });
   }